Create second page (input validation should be added)
On the first page, ask for the user name; after that, the page allows the user to choose the animal that the user wants to know about.
<Special Note>
When we use a form and wanna use a "button", make sure the type is "button". If not, it will submit the form right away.
<Additional Note>
Name validation function should be developed.